# packages/contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Montenegrina API
  version: 0.1.0
  description: Provider-independent Montenegrin conversational-AI control and data plane.
servers:
  - url: http://localhost:3001
security:
  - apiKey: []
  - sessionCookie: []
...
paths:
...
  /v1/auth/login:
    post:
      operationId: login
      tags: [Authentication]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200': { $ref: '#/components/responses/Session' }
        '401': { $ref: '#/components/responses/Error' }
...
  /v1/auth/me:
    get:
      operationId: getCurrentUser
      tags: [Authentication]
      responses:
        '200': { $ref: '#/components/responses/Session' }
        '401': { $ref: '#/components/responses/Error' }
...
  /v1/transcriptions:
    post:
      operationId: createTranscription
      tags: [Providers]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [audio]
              properties:
                audio: { type: string, format: binary }
                agentId: { type: string, format: uuid }
                sttLanguage: { $ref: '#/components/schemas/SttLanguage' }
      responses:
        '200': { $ref: '#/components/responses/Transcription' }
  /v1/responses:
    post:
      operationId: createResponse
      tags: [Providers]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateResponseRequest' }
      responses:
        '200': { $ref: '#/components/responses/GeneratedResponse' }
        '400': { $ref: '#/components/responses/Error' }
  /v1/speech:
    post:
      operationId: createSpeech
      tags: [Providers]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateSpeechRequest' }
      responses:
        '200':
          description: Streamed audio
          headers:
            X-Request-Id: { schema: { type: string } }
            X-Provider: { schema: { type: string } }
          content:
            audio/wav: { schema: { type: string, format: binary } }
            audio/pcm: { schema: { type: string, format: binary } }
  /v1/conversations:
    post:
      operationId: createConversation
      tags: [Conversations]
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [agentId, channel]
              properties:
                agentId: { type: string, format: uuid }
                channel: { type: string, enum: [TEXT, BROWSER, SIP, BATCH] }
      responses:
        '201': { $ref: '#/components/responses/Conversation' }
...
components:
  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: Authorization
    sessionCookie:
      type: apiKey
      in: cookie
      name: montenegrina_session
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 8, maxLength: 200 }
...
  schemas:
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, requestId, retryable]
          properties:
            code: { type: string }
            message: { type: string }
            requestId: { type: string }
            retryable: { type: boolean }
            details: { type: object, additionalProperties: true }
    Language:
      type: string
      const: cnr
...
    SttLanguage:
      type: string
      enum: [sr, hr, bs, multi]
    LoginRequest:
      type: object
      additionalProperties: false
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 12, maxLength: 200 }
    Session:
      type: object
      additionalProperties: false
      required: [user, csrfToken]
      properties:
        user: { $ref: '#/components/schemas/User' }
        csrfToken: { type: string }
    User:
      type: object
      additionalProperties: false
      required: [id, email, createdAt]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        displayName: { type: string }
        createdAt: { type: string, format: date-time }
...
    Transcription:
      type: object
      additionalProperties: false
      required: [language, originalText, normalizedText, warnings, provider]
      properties:
        language: { $ref: '#/components/schemas/Language' }
        originalText: { type: string }
        normalizedText: { type: string }
        warnings: { type: array, items: { type: string } }
        provider: { $ref: '#/components/schemas/ProviderSummary' }
    CreateResponseRequest:
      type: object
      additionalProperties: false
      required: [agentId, input]
      properties:
        agentId: { type: string, format: uuid }
        conversationId: { type: string, format: uuid }
        input: { type: string, minLength: 1, maxLength: 50000 }
        stream: { type: boolean, default: false }
    GeneratedResponse:
      type: object
      additionalProperties: false
      required: [language, text, originalText, warnings, citations, provider]
      properties:
        language: { $ref: '#/components/schemas/Language' }
        text: { type: string }
        originalText: { type: string }
        warnings: { type: array, items: { type: string } }
        citations: { type: array, items: { $ref: '#/components/schemas/Citation' } }
        provider: { $ref: '#/components/schemas/ProviderSummary' }
    CreateSpeechRequest:
      type: object
      additionalProperties: false
      required: [agentId, text]
      properties:
        agentId: { type: string, format: uuid }
        text: { type: string, minLength: 1, maxLength: 10000 }
        format: { type: string, enum: [wav, pcm], default: wav }
    ProviderSummary:
      type: object
      additionalProperties: false
      required: [provider, model, latencyMs, usage]
      properties:
        provider: { type: string }
        model: { type: string }
        latencyMs: { type: integer, minimum: 0 }
        usage: { $ref: '#/components/schemas/Usage' }
        fallbackHistory:
          type: array
          items:
            type: object
            required: [provider, code]
            properties:
              provider: { type: string }
              code: { type: string }
              retryable: { type: boolean }
    Usage:
      type: object
      additionalProperties: false
      properties:
        provider: { type: string }
        operation: { type: string }
        inputTokens: { type: integer, minimum: 0 }
        outputTokens: { type: integer, minimum: 0 }
        audioInputSeconds: { type: number, minimum: 0 }
        audioOutputSeconds: { type: number, minimum: 0 }
        characters: { type: integer, minimum: 0 }
        estimatedCostUsd: { type: number, minimum: 0 }
    Conversation:
      type: object
      additionalProperties: false
      required: [id, organizationId, agentId, channel, state, language, startedAt]
      properties:
        id: { type: string, format: uuid }
        organizationId: { type: string, format: uuid }
        agentId: { type: string, format: uuid }
        channel: { type: string, enum: [TEXT, BROWSER, SIP, BATCH] }
        state: { $ref: '#/components/schemas/ConversationState' }
        language: { $ref: '#/components/schemas/Language' }
        startedAt: { type: string, format: date-time }
        completedAt: { type: [string, 'null'], format: date-time }
    ConversationState:
      type: string
      enum: [INITIALIZING, LISTENING, TRANSCRIBING, THINKING, TOOL_PENDING, SPEAKING, INTERRUPTED, HANDOFF_PENDING, HANDED_OFF, COMPLETED, FAILED]
...
    Citation:
      type: object
      additionalProperties: false
      required: [documentId, documentVersionId, chunkId, title]
      properties:
        documentId: { type: string, format: uuid }
        documentVersionId: { type: string, format: uuid }
        chunkId: { type: string, format: uuid }
        title: { type: string }
        page: { type: integer, minimum: 1 }
        section: { type: string }
        score: { type: number }
...
  responses:
    Error:
      description: Typed error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
    Session:
      description: Browser session
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Session' }
...
    Transcription:
      description: Normalized transcription
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Transcription' }
    GeneratedResponse:
      description: Montenegrin response
      content:
        application/json:
          schema: { $ref: '#/components/schemas/GeneratedResponse' }
    Conversation:
      description: Conversation
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Conversation' }
...